#113 / Вынос создания diagnostic в helper
